+++ NEWS.md
@@ -5,7 +5,9 @@ PSFCalc 0.1.2
 * Fixed: Reference for LimitState library was not appropriate. 'LimitState::' is added in front of LSFM and Lbase.
 
 ### New features
+* Target command is added.
 * PSFEval command is added.
+* ShellRt command is added.
 * PSFContour command is added.
 
 PSFCalc 0.1.1